Refactor hyper tests
