+++ src/instrumentations/http/hyper/mod.rs
@@ -57,6 +57,9 @@ mod legacy_client;
 #[cfg(feature = "hyper-client-legacy")]
 pub use legacy_client::HyperLegacyClientInstrument;
 
+#[cfg(test)]
+mod test_utils;
+
 impl UrlParts for hyper::Uri {
     fn full_url(&self) -> Option<impl Into<Value>> {
         Some(self.to_string())
@@ -250,72 +253,25 @@ fn hyper_error_type(error: &hyper::Error) -> &'static str {
 
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "hyper-client-legacy")]
-    use super::HyperLegacyClientInstrument;
-
-    use super::HyperSendRequestInstrument;
-    use crate::{Context, Value, semconv};
+    use super::{HyperSendRequestInstrument, test_utils::*};
+    use crate::{Context, semconv};
     use assert2::assert;
-    use axum::{
-        Router,
-        extract::State,
-        http::{HeaderMap, StatusCode},
-        response::{IntoResponse, Redirect},
-        routing::get,
-    };
+    use axum::http::StatusCode;
     use bytes::Bytes;
     use http_body_util::Empty;
     use hyper::{
         Request,
         header::{HOST, USER_AGENT},
     };
-    #[cfg(any(feature = "hyper-http2", feature = "hyper-client-legacy"))]
+    #[cfg(feature = "hyper-http2")]
     use hyper_util::rt::TokioExecutor;
     use hyper_util::rt::TokioIo;
-    use opentelemetry::{
-        global,
-        trace::{Span as _, SpanKind, TraceContextExt, Tracer as _, TracerProvider as _},
-    };
-    use opentelemetry_sdk::{
-        propagation::TraceContextPropagator,
-        trace::{InMemorySpanExporter, SdkTracerProvider as TracerProvider},
+    use opentelemetry::trace::{
+        Span as _, SpanKind, TraceContextExt, Tracer as _, TracerProvider as _,
     };
     use serial_test::serial;
-    use std::sync::{Arc, Mutex};
-    use tokio::{
-        net::{TcpListener, TcpStream},
-        task::JoinHandle,
-    };
+    use tokio::net::TcpStream;
     use tracing_subscriber::{Registry, layer::SubscriberExt};
-
-    #[derive(Clone, Default)]
-    struct TestState {
-        traceparents: Arc<Mutex<Vec<(String, String)>>>,
-    }
-
-    impl TestState {
-        fn record(&self, path: &str, headers: &HeaderMap) {
-            if let Some(traceparent) = headers
-                .get("traceparent")
-                .and_then(|value| value.to_str().ok())
-            {
-                self.traceparents
-                    .lock()
-                    .unwrap()
-                    .push((path.to_owned(), traceparent.to_owned()));
-            }
-        }
-
-        fn traceparent_for(&self, path: &str) -> Option<String> {
-            self.traceparents
-                .lock()
-                .unwrap()
-                .iter()
-                .rev()
-                .find(|(recorded_path, _)| recorded_path == path)
-                .map(|(_, traceparent)| traceparent.clone())
-        }
-    }
 
     #[cfg(feature = "hyper-http1")]
     #[tokio::test]
@@ -520,302 +476,5 @@ mod tests {
         ));
         assert!(i64_attr(span, semconv::HTTP_RESPONSE_STATUS_CODE) == Some(404));
         assert!(string_attr(span, semconv::ERROR_TYPE) == Some("404"));
-    }
-
-    #[cfg(feature = "hyper-client-legacy")]
-    #[tokio::test]
-    #[serial]
-    async fn instruments_successful_legacy_client_requests() {
-        let telemetry = configure_test_tracing();
-        let server = spawn_server().await;
-        let client = hyper_util::client::legacy::Client::builder(TokioExecutor::new())
-            .build_http::<Empty<Bytes>>()
-            .instrument();
-        let request_url = format!("{}/ok?ready=true", server.base_url);
-        let response = client
-            .request(
-                Request::builder()
-                    .uri(&request_url)
-                    .header(USER_AGENT, "telemetry-rust-tests")
-                    .body(Empty::<Bytes>::new())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert!(response.status() == StatusCode::OK);
-
-        let spans = force_flush_and_get_spans(&telemetry);
-        let span = find_span(&spans, "GET");
-        let traceparent = server.state.traceparent_for("/ok").unwrap();
-        let (trace_id, span_id) = traceparent_ids(&traceparent);
-
-        assert!(span.span_kind == SpanKind::Client);
-        assert!(span.span_context.trace_id().to_string() == trace_id);
-        assert!(span.span_context.span_id().to_string() == span_id);
-        assert!(matches!(span.status, opentelemetry::trace::Status::Unset));
-        assert!(string_attr(span, semconv::HTTP_REQUEST_METHOD) == Some("GET"));
-        assert!(string_attr(span, semconv::URL_SCHEME) == Some("http"));
-        assert!(string_attr(span, semconv::SERVER_ADDRESS) == Some("127.0.0.1"));
-        assert!(
-            i64_attr(span, semconv::SERVER_PORT) == Some(i64::from(server.addr.port()))
-        );
-        assert!(string_attr(span, semconv::URL_PATH) == Some("/ok"));
-        assert!(string_attr(span, semconv::URL_QUERY) == Some("ready=true"));
-        assert!(
-            string_attr(span, semconv::USER_AGENT_ORIGINAL)
-                == Some("telemetry-rust-tests")
-        );
-        assert!(string_attr(span, semconv::URL_FULL) == Some(request_url.as_str()));
-        assert!(i64_attr(span, semconv::HTTP_RESPONSE_STATUS_CODE) == Some(200));
-        assert!(string_attr(span, semconv::NETWORK_PROTOCOL_VERSION).is_some());
-    }
-
-    #[cfg(feature = "hyper-client-legacy")]
-    #[tokio::test]
-    #[serial]
-    async fn legacy_client_uses_explicit_parent_context_when_provided() {
-        let telemetry = configure_test_tracing();
-        let server = spawn_server().await;
-        let client = hyper_util::client::legacy::Client::builder(TokioExecutor::new())
-            .build_http::<Empty<Bytes>>();
-        let tracer = telemetry.provider.tracer("hyper-legacy-tests");
-        let explicit_parent = tracer.start("explicit-parent");
-        let explicit_parent_span_id = explicit_parent.span_context().span_id();
-        let explicit_parent_cx = Context::current_with_span(explicit_parent);
-        let tracing_tracer = telemetry.provider.tracer("tracing-tests");
-        let subscriber = Registry::default()
-            .with(tracing_opentelemetry::layer().with_tracer(tracing_tracer));
-        let guard = tracing::subscriber::set_default(subscriber);
-        let current_parent = tracing::info_span!("current-parent");
-
-        tracing::Instrument::instrument(
-            async {
-                let client = client.instrument().context(&explicit_parent_cx);
-                client
-                    .request(
-                        Request::builder()
-                            .uri(format!("{}/ok", server.base_url))
-                            .body(Empty::<Bytes>::new())
-                            .unwrap(),
-                    )
-                    .await
-                    .unwrap();
-            },
-            current_parent,
-        )
-        .await;
-
-        drop(guard);
-        explicit_parent_cx.span().end();
-
-        let spans = force_flush_and_get_spans(&telemetry);
-        let hyper_span = find_span(&spans, "GET");
-        let current_span = find_span(&spans, "current-parent");
-
-        assert!(hyper_span.parent_span_id == explicit_parent_span_id);
-        assert!(hyper_span.parent_span_id != current_span.span_context.span_id());
-    }
-
-    #[cfg(feature = "hyper-client-legacy")]
-    #[tokio::test]
-    #[serial]
-    async fn legacy_client_marks_error_responses_as_errors() {
-        let telemetry = configure_test_tracing();
-        let server = spawn_server().await;
-        let client = hyper_util::client::legacy::Client::builder(TokioExecutor::new())
-            .build_http::<Empty<Bytes>>()
-            .instrument();
-        let response = client
-            .request(
-                Request::builder()
-                    .uri(format!("{}/not-found", server.base_url))
-                    .body(Empty::<Bytes>::new())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert!(response.status() == StatusCode::NOT_FOUND);
-
-        let spans = force_flush_and_get_spans(&telemetry);
-        let span = find_span(&spans, "GET");
-
-        assert!(matches!(
-            span.status,
-            opentelemetry::trace::Status::Error { .. }
-        ));
-        assert!(i64_attr(span, semconv::HTTP_RESPONSE_STATUS_CODE) == Some(404));
-        assert!(string_attr(span, semconv::ERROR_TYPE) == Some("404"));
-    }
-
-    #[cfg(feature = "hyper-client-legacy")]
-    #[tokio::test]
-    #[serial]
-    async fn legacy_client_marks_transport_errors_as_errors() {
-        let telemetry = configure_test_tracing();
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        drop(listener);
-        let client = hyper_util::client::legacy::Client::builder(TokioExecutor::new())
-            .build_http::<Empty<Bytes>>()
-            .instrument();
-        let result = client
-            .request(
-                Request::builder()
-                    .uri(format!("http://{addr}/unreachable"))
-                    .body(Empty::<Bytes>::new())
-                    .unwrap(),
-            )
-            .await;
-
-        assert!(result.is_err());
-
-        let spans = force_flush_and_get_spans(&telemetry);
-        let span = find_span(&spans, "GET");
-
-        assert!(matches!(
-            span.status,
-            opentelemetry::trace::Status::Error { .. }
-        ));
-        assert!(string_attr(span, semconv::ERROR_TYPE) == Some("connect"));
-    }
-
-    struct TestServer {
-        addr: std::net::SocketAddr,
-        base_url: String,
-        state: TestState,
-        _handle: JoinHandle<()>,
-    }
-
-    impl TestServer {
-        fn authority(&self) -> String {
-            self.addr.to_string()
-        }
-    }
-
-    async fn spawn_server() -> TestServer {
-        async fn ok(
-            State(state): State<TestState>,
-            headers: HeaderMap,
-        ) -> impl IntoResponse {
-            state.record("/ok", &headers);
-            StatusCode::OK
-        }
-
-        async fn not_found(
-            State(state): State<TestState>,
-            headers: HeaderMap,
-        ) -> impl IntoResponse {
-            state.record("/not-found", &headers);
-            StatusCode::NOT_FOUND
-        }
-
-        async fn redirect(
-            State(state): State<TestState>,
-            headers: HeaderMap,
-        ) -> impl IntoResponse {
-            state.record("/redirect", &headers);
-            Redirect::temporary("/final")
-        }
-
-        async fn final_route(
-            State(state): State<TestState>,
-            headers: HeaderMap,
-        ) -> impl IntoResponse {
-            state.record("/final", &headers);
-            StatusCode::OK
-        }
-
-        let state = TestState::default();
-        let app = Router::new()
-            .route("/ok", get(ok))
-            .route("/not-found", get(not_found))
-            .route("/redirect", get(redirect))
-            .route("/final", get(final_route))
-            .with_state(state.clone());
-
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let handle = tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-
-        TestServer {
-            addr,
-            base_url: format!("http://{addr}"),
-            state,
-            _handle: handle,
-        }
-    }
-
-    fn configure_test_tracing() -> TestTelemetry {
-        let exporter = InMemorySpanExporter::default();
-        let provider = TracerProvider::builder()
-            .with_simple_exporter(exporter.clone())
-            .build();
-        global::set_tracer_provider(provider.clone());
-        global::set_text_map_propagator(TraceContextPropagator::new());
-        TestTelemetry { exporter, provider }
-    }
-
-    fn force_flush_and_get_spans(
-        telemetry: &TestTelemetry,
-    ) -> Vec<opentelemetry_sdk::trace::SpanData> {
-        telemetry.provider.force_flush().unwrap();
-        telemetry.exporter.get_finished_spans().unwrap()
-    }
-
-    fn find_span<'a>(
-        spans: &'a [opentelemetry_sdk::trace::SpanData],
-        name: &str,
-    ) -> &'a opentelemetry_sdk::trace::SpanData {
-        spans.iter().find(|span| span.name == name).unwrap()
-    }
-
-    fn string_attr<'a>(
-        span: &'a opentelemetry_sdk::trace::SpanData,
-        key: &str,
-    ) -> Option<&'a str> {
-        match attr(span, key) {
-            Some(Value::String(value)) => Some(value.as_str()),
-            _ => None,
-        }
-    }
-
-    fn i64_attr(span: &opentelemetry_sdk::trace::SpanData, key: &str) -> Option<i64> {
-        match attr(span, key) {
-            Some(Value::I64(value)) => Some(*value),
-            _ => None,
-        }
-    }
-
-    fn attr<'a>(
-        span: &'a opentelemetry_sdk::trace::SpanData,
-        key: &str,
-    ) -> Option<&'a Value> {
-        span.attributes
-            .iter()
-            .find(|kv| kv.key.as_str() == key)
-            .map(|kv| &kv.value)
-    }
-
-    fn traceparent_ids(traceparent: &str) -> (&str, &str) {
-        let mut parts = traceparent.split('-');
-        let _version = parts.next().unwrap();
-        let trace_id = parts.next().unwrap();
-        let span_id = parts.next().unwrap();
-        (trace_id, span_id)
-    }
-
-    struct TestTelemetry {
-        exporter: InMemorySpanExporter,
-        provider: TracerProvider,
-    }
-
-    impl Drop for TestTelemetry {
-        fn drop(&mut self) {
-            let _ = self.provider.shutdown();
-        }
     }
 }
